+++ www/routes/businessLogic/render.js
@@ -116,6 +116,7 @@ router.post('/list', checkIfLoggedIn, function (req, res, next) {
     promise_array.push(s3.list_folders(req.user.email + '/GraphQL/youtube-Search-Playlist/'));
     promise_array.push(s3.list_folders(req.user.email + '/GraphQL/youtube-Most-Popular/'));
     promise_array.push(s3.list_folders(req.user.email + '/GraphQL/youtube-Creator-Videos/'));
+    promise_array.push(s3.list_folders(req.user.email + '/GraphQL/youtube-Random-Videos/'));
     promise_array.push(s3.list_folders(req.user.email + '/GraphQL/userspec-Others/'));
     Promise.all(promise_array).then(values => {
         directory['twitter-Tweet'] = values[0];
@@ -131,7 +132,8 @@ router.post('/list', checkIfLoggedIn, function (req, res, next) {
         directory['youtube-Search-Playlist'] = values[10];
         directory['youtube-Most-Popular'] = values[11];
         directory['youtube-Creator-Videos'] = values[12];
-        directory['userspec-Others'] = values[13];
+        directory['youtube-Random-Videos'] = values[13];
+        directory['userspec-Others'] = values[14];
         res.send(directory);
     }).catch(err => {
         res.send({ERROR: err});
@@ -154,6 +156,7 @@ router.post('/list-all', checkIfLoggedIn, function (req, res, next) {
     promise_array.push(s3.list_folders(req.user.email + '/GraphQL/youtube-Search-Playlist/'));
     promise_array.push(s3.list_folders(req.user.email + '/GraphQL/youtube-Most-Popular/'));
     promise_array.push(s3.list_folders(req.user.email + '/GraphQL/youtube-Creator-Videos/'));
+    promise_array.push(s3.list_folders(req.user.email + '/GraphQL/youtube-Random-Videos/'));
     promise_array.push(s3.list_folders(req.user.email + '/GraphQL/userspec-Others/'));
     var graphqlLength = promise_array.length;
 
@@ -193,7 +196,8 @@ router.post('/list-all', checkIfLoggedIn, function (req, res, next) {
         directory['GraphQL']['youtube-Search-Playlist'] = values[10];
         directory['GraphQL']['youtube-Most-Popular'] = values[11];
         directory['GraphQL']['youtube-Creator-Videos'] = values[12];
-        directory['GraphQL']['userspec-Others'] = values[13];
+        directory['GraphQL']['youtube-Random-Videos'] = values[13];
+        directory['GraphQL']['userspec-Others'] = values[14];
         for (i = 0; i < order.length; i++) {
             var parent = order[i]['parent'];
             var child = order[i]['child'];