+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Randomly collect youtube video [#169](https://github.com/ncsa/standalone-smm-smile/issues/169)
 
 ## [0.4.0] - 2024-06-13
 ### Added